+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- 2016/09/20 (1.154)- Fixed color button size calculation
 - 2016/09/20 (1.153)- Fixed some `nk_vsnprintf` behavior bugs and removed
                         `<stdio.h>` again from `NK_INCLUDE_STANDARD_VARARGS`.
 - 2016/09/18 (1.152)- C89 does not support vsnprintf only C99 and newer as well